+++ sdk/mongocluster/arm-mongocluster/src/classic/index.ts
@@ -7,3 +7,4 @@ export { OperationsOperations } from "./operations/index.js";
 export { PrivateEndpointConnectionsOperations } from "./privateEndpointConnections/index.js";
 export { PrivateLinksOperations } from "./privateLinks/index.js";
 export { ReplicasOperations } from "./replicas/index.js";
+export { UsersOperations } from "./users/index.js";