+++ sdk/ai/ai-projects/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - The `project.datasets.uploadFile` method is updated to take an options bag of type `DatasetUploadOptionalOptions`.
 - The `project.datasets.uploadFolder` method is updated to take an options bag of type `DatasetUploadOptionalOptions`.
+- Add the apiVersion and rest api doc information of client library to `README.md`
 
 ## 1.0.0-beta.8 (2025-05-21)